Split the inference, sensitive column handling, and SQL query execution.
This is a pure refactor, but in the following change I plan to move the SQL query execution to SessionCommit as a step toward custom DP units where the SQL query is executed at the DP unit scope. Inference and sensitive column handling, which have a row-level scope, will remain in SessionAccumulate.

Change-Id: Ia380ee2403ea9bfe5f424f084ae491edbf5f5d2e