Updated Confidential Transoform Session to allow writing an arbitrary number of ReadResponse messages for any request.
This introduces a new Session interface with an ability to write back an arbitrary number of blobs.
For backwards compatibility, this also add a LegacySession class that works as a shim - it
 provides implementation of the new Session interface and surfaces a set of virtual methods that
match the old Session interface.
All of the existing Session implementations have been changed to derive from LegacySession.
We can move them to the new interface as necessary.

Change-Id: I325167249726f27ff56880471b32cb80dda159d6