Use a single Rust function to create SessionConfig for program worker.
This CL moves the session config creation logic to Rust to make it
easier to use other implementations of attester/endorser and session
binders.

The C API function will receive the endorsed evidence as a byte array
and then use it to create a `SessionConfig` which will then be returned
as a raw pointer. The caller will be responsible for managing the memory
of this object.

TESTED: manually built the binary and tested it internally
Change-Id: Ie2e290ddb6b72defda8e297a661cfe1e99305e8e