docs: remove table from KMS readme
The Mermaid graph is rendered using the available space, which is very
small since nothing is forcing the table cell to grow. This results in
an unreadably small graph:

https://github.com/google-parfait/confidential-federated-compute/tree/b62ff5ccefc1df25f508275bf108147545d92985/containers/kms#access-policies

Since GitHub's sanitization strips style tags, we can't use `min-width`
or `height`. Instead, we abandon the two-column rendering; this matches
GitHub's Mermaid examples:
https://docs.github.com/en/get-started/writing-on-github/working-with-advanced-formatting/creating-diagrams.

Bug: 392128673
Change-Id: Idad2d02e930b19c0f3aa3991e2747c6df4e09532