Add caching for Data uris to the execution context.
Each time we make a request to untrusted space to resolve a Data uri
we subtract budget. Regardless of how many times a client upload is
used within a program instance, we only want to subtract budget once.
For now, this means that we need to hold a resolved piece of data in
the program executor TEE memory after is it is resolve the first time.
In the future when we switch to the reduced state ledger this strategy
will likely change.

Change-Id: I1ff4b7655d43f7cd038752949bf879df89909f73