Create a minimum container that supports jax and python
This container installs jax and sentencepiece python libraries and the
entire data processing is done in Python. The main entry point to the
container is still a cc binary. Pybind is used for in-process cc-python
communication. This container also shows one simple example to
processing string data with the combination of sentencepiece and jax.

This container doesn't support CUDA yet.
Confidential transform API will be added in a subsequent change.

Bug: 439886759
Change-Id: Iddb17eaf87d2b13034445bf35b7a61d97c073ed1