Add RowSet and RowView classes.
RowSet is a non-owning view of a set of rows from a collection of
uploads. The rows in the set is specified by an absl::Span<RowLocation>.
The FedSQL container will specify the absl::Span<RowLocation> such that
rows are grouped by DP unit.

RowView is a non-owning view of a single row. We implement an STL
iterator for RowView so it's more convenient to access the elements of
the row. We'll update SqliteAdapter to use RowSet and RowView.
SqliteAdapter::InsertRows will be updated to iterate over RowSet and
RowView instead of the current double for loop over indices.

BUG: 432091990
Change-Id: Ifb998cc4c47b47e541711cfdd5cabbcfa398ee9b