Refactor uncommitted input storage in KmsFedSqlSession.
The `uncommitted_inputs_` member is changed from a map to a vector, and
a separate `uncommitted_blob_ids_` set is used to track unique blob IDs.

This is a small change, since @edor is working on the SessionAccumulate
changes and will be using the new members added by this change.

In later changes I will add a row_dp_unit_index_ vector, update
SessionCommit to sort it, and execute the SQL query at the DP unit
scope.

BUG: 432091990
Change-Id: I4c0d3bc57a7615a31e10b2e616f891ebf73196b5