Add `dp_unit` library for processing SQL queries grouped by DP units.
This CL introduces the `dp_unit` library, which provides the
`CommitRowsGroupingByDpUnit` function. This function takes a set of
uncommitted SQL inputs and a list of `RowLocation`s, which map rows to
DP units. For each DP unit, it executes the provided SQL query on the
subset of rows belonging to that DP unit. The results of each
per-DP-unit query execution are then accumulated into a
`CheckpointAggregator`. A `<` operator is added to `RowLocation` to
support sorting by `dp_unit_hash`, `input_index`, and `row_index`.

The new library allows us to test CommitRowsGroupingByDpUnit without
going through the whole session flow (ie. a few accumulates, commit,
finally serialize or report). It also allows me to submit code while Edo
is working on the DP unit changes in the Accumulate method.

I will call this from KmsFedSqlSession::Commit once Edo's changes to
Accumulate are submitted.

BUG:432091990
Change-Id: I0744962f04adb0ef31a2f9202b573eeabb5a62e2