Move TF program executor code to a separate directory.
This is one change in a sequence that prepares for adding both a simple
python version and a JAX version of the program executor container in
learning_containers/program_executor_tee.

In the next change the code in the learning_containers/tf/ directory
will be moved to a new learning_tensorflow_containers/ directory with
its own WORKSPACE file. Because this new WORKSPACE file will be pinned
to a specific version of CFC and the learning_tensorflow_containers/
code will depend on code in learning_containers/, we need two separate
cls (this cl, which is separating "base" logic and "TF" logic in
learning_containers/, and a follow-up cl adding
learning_tensorflow_containers/).

The confidential_transform_server class in learning_containers/
program_executor_tee can now function as a base class with a
GetProgramInitializeFn function that can be overridden by subclasses.
This function is passed to the program_runner and is run prior to
executing the customer-provided program. Integration tests within
program_runner_test.py that used to test the execution stack are removed
and program_runner_test.py instead functions now as a unit test. The
same version of program_runner.py should be used by all containers and
it should not need to be forked.

FakeComputationDelegationService now takes an arbitrary leaf executor
factory in the constructor (as opposed to always using a TF executor),
and FakeComputationDelegationServiceTest uses XLA instead of TF. There
are now separate fake_service_bindings modules for JAX and TF.

Change-Id: Ibfb51e4a82d6579fe79f58ac3b153ea575268536