Update CFC version in learning_containers WORKSPACE.
The libcppbor patch (which is no longer present in CFC) is needed
by the learning_containers, so it is copied over.

A protobuf patch is added since the version of protobuf used by
learning_containers (because of TF) is now very different from
the version of protobuf used in the top-level CFC WORKSPACE.

This cl also makes a few small changes to migrate the program
executor TEE's confidential transform implementation to the new
session API. The benchmarking tff server confidential transform
implementation is not migrated (but rather the LegacySession
class is used) since we anticipate removing the tff server soon.

Change-Id: Ib6dc927e419f6024044dee2e8f4f0e6f01d616ba