Adjust TFF compilation/execution stack in distributed case.
Instead of using to_composed_tee_form and the ComposingTeeExecutor/
TeeExecutor/LambdaRunner, use to_call_dominant_form and
ComposingExecutor/RemoteExecutor/ExecutorService. The former approach
was built specifically for our initial benchmarking infra but is not
as generic/widely tested as the latter approach, which will be an
easier approach to maintain going forward.

Two main new classes are added: NoiseExecutorStub and NoiseLeafExecutor.
NoiseExecutorStub wraps a NoiseClientSession and is provided as input
to a RemoteExecutor. It takes the Executor proto messages the
RemoteExecutor is trying to send/receive and wraps them into a form
that can be transmitted using the NoiseClientSession. The
NoiseLeafExecutor wraps a oak ServerSession, obtains the wrapped
Executor proto messages and uses a ExecutorService to perform the leaf
operations. The NoiseLeafExecutor is used both in root executor tests
(it is used by the FakeComputationDelegationService) and by the
program worker binary.

This change also consolidates the program executor TEE and program
worker code together in learning_containers/program_executor_tee. Note
that we will need to add back in TF program worker code in a separate
cl (this code now belongs in the learning_tensorflow_containers
directory, and the WORKSPACE file in that directory will need to update
the version of CFC to include this cl before those changes can be made).

Change-Id: I846c9398cc53ff08e07f884e8f0f15124c73e5ce