fix: Add a timeout to StorageClient initialization
We've seen occasional hangs during initialization where the process
starts but does not complete. To prevent these hangs from being
permanent, we add a 1 minute deadline to the initialization process --
which is significantly longer than the second or two it normally takes.
On timeout, a warning will be logged and initialization will be
re-attempted.

Bug: 448186987
Change-Id: I296db63b8de9f167bf65d9430cb1397375ab5d73