Add Oak to gcp_prototype container.
Oak doesn't fully support bzlmod yet, and we have to create it at the
root level in order to provide the dependencies from root.
Currently, rules_rust doesn't support module well which blocks oak to
fully support bzlmod.  More details in: https://github.com/bazelbuild/rules_rust/issues/3521

We need to use macros and toolchains that are defined inside oak repo,
it might be possible to write multiple extensions, but that'll be too
much work. To simplify the setup, we enable WORKSPACE to take care of
the parts that are not compatible with bzlmod. Once rules_rust and oak
support bzlmod properly, we'll delete the WORKSPACE file.  This also
means we can't migrate to bazel 9 which doesn't support WORKSPACE
anymore.

Change-Id: I3605fe6e257c517fe3ae3ba4ae83d8e4b6d8f0ef