Make the FedSQL container handle Message-based checkpoints.
If the init config contains the logged message file descriptor set and
message name, create a prototype Message and pass it to Sessions.
Sessions use the prototype to parse serialized messages from input
protos.

Inference and hashing sensitive columns is still not supported for
Message-based checkpoints. In a following CL I hope to refactor those to
operate on RowViews instead of Tensors so we can use the same
implementation for both types of checkpoints.

BUG:451799471

Change-Id: Ic56031047f02afa71dcaa1d076d247d7fa8f53d9