Update SqliteAdapter to use table schema column names.
In some cases the RowSet can contain columns that aren't present in the
table schema. For example, columns used as input for inference are not
present in the table schema. This is not a problem for Tensor-based
Inputs since our inference logic actually removes those columns.
However, Message-based Inputs won't be able to remove those columns if
they're fields in the logged proto.

Instead, AddTableContents finds the columns in the RowSet that are also
in the table schema and only inserts those columns.

In my next change I plan to refactor RunInference to operate on Inputs
instead of std::vector<Tensor> so it can handle both Input types.

BUG:451799471
Change-Id: I6deebaaa77f6372b2a1dc17d326e0e71bdf00b68