Refactor InferenceModel to use Input.
RunInference now takes an Input instead of a vector<Tensor>. It also has
been modified to NOT remove the inference input columns, since
Message-based inputs can't remove input columns that are part of the
logged message. The SqliteAdapter::AddTableContents was update in a
previous change to handle this.

Also add an AddColumn method to Input so we can add the inference output
column.

I refactored a lot of the unit tests - I added a test fixture and
instead of having logic for the mock RunGemmaCppInference it now returns
a set output Tensor. I think we should refactor InferenceModel to
decrease the amount we have to mock - currently we can't actually test
some of the logic in RunGemmaCppInference since we're mocking it.
However, this CL is already large so I didn't include that refactoring.
I removed a couple tests that were testing logic in RunGemmaCppInference
since they weren't actually testing anything meaningful.

BUG:451799471
Change-Id: Iee2941adecb7b883504398c8321d8da8028120d7