Update budget tracking for interval overlap scenario.
Say we have the following scenario:
Total interval: [0, 10), budget 1, consumed_range = [0, 4)

Now say we consume [3, 10) - meaning we overlap with the consumed_range.

Previously, if there was an overlap, we would carryover the full interval that caused the overlap. Meaning we would then have the following: Budget 0, consumed_range = [3, 10).

However, this overconsumes the budget (it was only 3 that was consumed beyond are partial budget). With this CL we would now end up with the following after consuming [3, 10): Budget 0, consumed_range [3, 4) (the intersection of the overlapping ranges).

Change-Id: I1801fd4a638adfc82163a3aabc4e7ba33975e367