Support uri resolution within DataParser.
The new ResolveUriToTensor function is not used yet, however in
follow-up cls it will be provided to the ProgramInputProvider
constructor so that data resolution is no longer hidden within the
TrustedExecutionContext (which non-TFF programs will not have).

While we are still using the ledger, caching is turned on by
default. When we migrate to KMS, we will no longer need to use
caching.

Change-Id: Idf52d8df3de086564c0fcd4038f2309569127ca8