Add gen_python_path rule.
This rule is designed to create PYTHONPATH for cc binaries calling
embeded python libraries. This rule only works for python deps introduced
via bzlmod, and it's not backward compatible with WORKSPACE

This rule is copied from the one inside minimum_jax container.  It is
promoted to the container/bazel directory to allow other containers
depend on it more easily. The one inside minimum_jax container will be
deleted in a later clean-up change after this one is in, and a new CFC
version is cut for bzlmod.

BUG:454053045

Change-Id: I35d4c43585af6776f24a07fd8f9f0da892c7e673