Adopt new federated-compute version.
As part of adopting the new ProgramInputProvider, we refactor
where data resolution happens in the root. The execution context
no longer resolves data. Instead the ProgramInputProvider is
constructed with a data resolution fn. Data sources can wrap this
ProgramInputProvider if they want to produce resolved data.

The ReleaseManager is renamed to ExternalServiceHandle and now
extends federated-compute's similarly named interface. Due to
this adoption, the trusted_program method in a customer-provided
program no longer needs to be async.

Bug: 457844838

Change-Id: Id46b8adef8e65bd1420e9e92ea21ce10bb56c294