Implement MetadataSession for extracting payload metadata.
This change introduces the MetadataSession, which processes unencrypted
data from a Federated Compute checkpoint. It extracts the privacy ID to
compute a partition key and event times to determine a coarse event time
range, based on the provided MetadataContainerConfig. The resulting
metadata is emitted as an unencrypted PayloadMetadataSet.

In a following change I'll implement
MetadataConfidentialTransform::CreateSession and add tests for Writes to
the MetadataConfidentialTransform server.

BUG:460521671

Change-Id: Iaee9105532ff45d9e53147ce31a6e943711a5e4b