Add a budget proto and private state class for the KMS integration.
For this initial implementation a very simple budget is used. The
budget just tracks the number of remaining runs for any program for
a given logical pipeline name. Because each release operation still
needs to trigger a state change, the budget also includes a
monotonically increasing counter.

In subsequent cls a PrivateState instance will be constructed within
the StreamInitializeTransformWithKms method, and the release logic
will access the PrivateState instance to generate release tokens
based on the current and next budget.

Bug: 451714072
Change-Id: I0b372693346cc230b8724d76660c6dc1bad6c629