Add support for auto-parsing LLM output in private inference.
This change introduces an "auto-parse" feature for LLM inference in FedSQL. When the prompt parser is set to PARSER_AUTO, the system appends instructions to the prompt, guiding the LLM to generate output in a JSON format. The JSON object will have a key matching the output column name and a value that is a JSON array. This allows for easier programmatic parsing of the LLM's response. The federated-compute dependency is also updated.

Change-Id: I3c5d130eeab050f221805a5441bf4ce15d0a2ece