Make FedSQL container handle column name prefixes.
- Pass the on-device query name to the KmsFedSqlSession
- Prefix the entry and event time column name constants with the on-device query name + "/" when reading a PrivateLogger upload
- Strip the on-device query name + "/" prefix from the event time Tensor, since SQL queries/schemas for pipelines that process PrivateLogger uploads don't prefix their column names with the on-device query name
- Update the test utils to generate more realistic Message-based checkpoints that include a tensor name prefix

We need this because on the client we prefix the PrivateLogger's entry
and event time tensors with the on-device query name to match the FedSQL
stack. I've already updated our TrustedAnalyticsWorkflow building logic
so it doesn't add this prefix to the table schema and SQL query for
pipelines that process PrivateLogger uploads.

I've also bumped the tensorflow_federated version to include a recent
change that allows us to rename Tensors.

More context can be found on the bug.

BUG:451799471
Change-Id: Iab5510bff724c508b7bf1ede2607e15d1b1df7b6