Refactor inference model to support multiple output rows per input row.
This change introduces an `InferenceOutput` struct to return both the
output tensor and the number of output rows generated per input row. A
new helper function, `duplicateColumnsForMultipleRows`, is added to
handle the cartesian product of multiple outputs by duplicating rows in
both the original input columns and the generated output columns (once
we support this). The inference output processing is also updated to
support number and boolean types within JSON arrays, and we also add a
test in inference_model to check auto-parsing with JSON.

Change-Id: Ie16cc911ac66f0feaec93eb7c407402ddffe9ee8