Merge expired keys in the RangeTracker.
If a key expired while partial aggregates are being computed, it's
possible that expired keys in different partitions will look different.
This is more likely to happen if a pipeline runs for more than 24 hours.
This change merges the expired keys instead of expecting all partitions
to have the same expired keys.


Change-Id: I0de064a7011f29b738f08a03117680c0721d5113